Add dictionary size parameter for LZW compression
The dictionary size can now be given an argument when performing LZW
compression. The dictionary size will be encoded into the compressed
file, so when decompressing, we don't need to know the dictionary size
used when compressing beforehand.

Tests were updated to reflect this change. They will be refactored
later.